Sumitra's Project
This project was created to teach people about living without healthcare. It is an interactive game which takes you through a scenario where you have to make decisions depending on whether or not you have healthcare. At the end of the simulation, the player is shown some real life statistics based on healthcare affordability and accessibility.